+++ brewtils/test/fixtures.py
@@ -348,9 +348,11 @@ def child_request_dict(ts_epoch):
         "status": "CREATED",
         "hidden": True,
         "command_type": "ACTION",
+        "root_command_type": "ACTION",
         "created_at": ts_epoch,
         "updated_at": ts_epoch,
         "status_updated_at": ts_epoch,
+        "expiration_at": ts_epoch,
         "error_class": None,
         "metadata": {"child": "stuff"},
         "has_parent": True,
@@ -367,6 +369,8 @@ def child_request(child_request_dict, ts_dt):
     dict_copy["created_at"] = ts_dt
     dict_copy["updated_at"] = ts_dt
     dict_copy["status_updated_at"] = ts_dt
+    dict_copy["expiration_at"] = ts_dt
+
     return Request(**dict_copy)
 
 
@@ -389,10 +393,12 @@ def parent_request_dict(ts_epoch):
         "output_type": "STRING",
         "status": "CREATED",
         "command_type": "ACTION",
+        "root_command_type": "ACTION",
         "created_at": ts_epoch,
         "hidden": False,
         "updated_at": ts_epoch,
         "status_updated_at": ts_epoch,
+        "expiration_at": ts_epoch,
         "error_class": None,
         "metadata": {"parent": "stuff"},
         "has_parent": False,
@@ -409,6 +415,7 @@ def parent_request(parent_request_dict, ts_dt):
     dict_copy["created_at"] = ts_dt
     dict_copy["updated_at"] = ts_dt
     dict_copy["status_updated_at"] = ts_dt
+    dict_copy["expiration_at"] = ts_dt
     return Request(**dict_copy)
 
 
@@ -457,9 +464,11 @@ def request_dict(parent_request_dict, child_request_dict, ts_epoch):
         "status": "CREATED",
         "hidden": False,
         "command_type": "ACTION",
+        "root_command_type": "ACTION",
         "created_at": ts_epoch,
         "updated_at": ts_epoch,
         "status_updated_at": ts_epoch,
+        "expiration_at": ts_epoch,
         "error_class": "ValueError",
         "metadata": {"request": "stuff"},
         "has_parent": True,
@@ -478,6 +487,7 @@ def bg_request(request_dict, parent_request, child_request, ts_dt):
     dict_copy["created_at"] = ts_dt
     dict_copy["updated_at"] = ts_dt
     dict_copy["status_updated_at"] = ts_dt
+    dict_copy["expiration_at"] = ts_dt
     return Request(**dict_copy)
 
 
@@ -1010,16 +1020,16 @@ def bg_status_info(status_info_dict, ts_dt, bg_status_history):
 
 @pytest.fixture
 def garden_dict(
-    ts_epoch, system_dict, connection_dict, connection_publishing_dict, status_info_dict
+    ts_epoch,
+    system_dict,
+    connection_dict,
+    connection_publishing_dict,
 ):
     """A garden as a dictionary."""
 
     return {
         "id": "123f11af55a38e64799fa1c1",
         "name": "garden",
-        "status": "RUNNING",
-        "status_info": status_info_dict,
-        "namespaces": [system_dict["namespace"]],
         "systems": [system_dict],
         "connection_type": "http",
         "receiving_connections": [connection_dict],
@@ -1036,14 +1046,16 @@ def garden_dict(
 
 @pytest.fixture
 def bg_garden(
-    garden_dict, bg_system, bg_connection, bg_connection_publishing, bg_status_info
+    garden_dict,
+    bg_system,
+    bg_connection,
+    bg_connection_publishing,
 ):
     """An operation as a model."""
     dict_copy = copy.deepcopy(garden_dict)
     dict_copy["systems"] = [bg_system]
     dict_copy["receiving_connections"] = [bg_connection]
     dict_copy["publishing_connections"] = [bg_connection_publishing]
-    dict_copy["status_info"] = copy.deepcopy(bg_status_info)
     return Garden(**dict_copy)